+++ app/models/stash_engine/user.rb
@@ -27,6 +27,7 @@
 #
 module StashEngine
   class User < ApplicationRecord
+    include StashEngine::Support::UserRoles
     self.table_name = 'stash_engine_users'
     has_many :roles, dependent: :destroy
     has_many :resources, through: :roles, source: :role_object, source_type: 'StashEngine::Resource'
@@ -46,12 +47,6 @@ module StashEngine
     has_many :access_tokens, through: :api_application, class_name: 'Doorkeeper::AccessToken'
 
     accepts_nested_attributes_for :roles, :flag
-
-    scope :curators, -> { joins(:roles).where('stash_engine_roles' => { role: 'curator', role_object_id: nil }) }
-
-    scope :all_curators,  -> { joins(:roles).where('stash_engine_roles' => { role: 'curator' }) }
-
-    scope :min_curators, -> { joins(:roles).where('stash_engine_roles' => { role: %w[superuser manager curator] }) }
 
     validates :email, format: { with: EMAIL_REGEX, message: '%{value} is not a valid email address' }, allow_blank: true
 
@@ -96,46 +91,6 @@ module StashEngine
 
     def name_last_first
       [last_name, first_name].join(', ')
-    end
-
-    def tenant_limited?
-      roles.any? { |r| r.role_object_type == 'StashEngine::Tenant' }
-    end
-
-    def admin?
-      roles.any? { |r| r.role == 'admin' }
-    end
-
-    def curator?
-      roles.any? { |r| r.role == 'curator' }
-    end
-
-    def manager?
-      roles.any? { |r| r.role == 'manager' }
-    end
-
-    def superuser?
-      roles.any? { |r| r.role == 'superuser' }
-    end
-
-    def system_user?
-      roles.any? { |r| r.role_object_id.nil? }
-    end
-
-    def min_admin?
-      roles.any? { |r| %w[superuser manager curator admin].include?(r.role) }
-    end
-
-    def min_app_admin?
-      system_user? || min_curator?
-    end
-
-    def min_curator?
-      roles.any? { |r| %w[superuser manager curator].include?(r.role) }
-    end
-
-    def min_manager?
-      roles.any? { |r| %w[superuser manager].include?(r.role) }
     end
 
     def proxy_user? = false
